feature/#34 - summary
