+++ Streamlit/Book_demo/aladin_api.py
@@ -12,13 +12,3 @@ url = f"http://www.aladin.co.kr/ttb/api/ItemSearch.aspx?ttbkey={TTBKey}&Query={q
 # API 요청
 response = requests.get(url)
 print(response.text)
-
-# # 응답 처리
-# if response.status_code == 200:
-#     data = response.json()
-#     for item in data['item']:
-#         title = item['title']
-#         description = item.get('description', '줄거리 없음')
-#         print(f"제목: {title}\n줄거리: {description}\n")
-# else:
-#     print(f"API 요청 실패: {response.status_code}")
